+++ tests/testthat/test-sym.R
@@ -27,6 +27,11 @@ test_that("is_symbol() matches `name`", {
   expect_false(is_symbol(sym("foo"), "bar"))
 })
 
+test_that("is_symbol() matches any name in a vector", {
+  expect_false(is_symbol(quote(C), letters))
+  expect_true(is_symbol(quote(c), letters))
+})
+
 test_that("must supply strings to sym()", {
   expect_error(sym(letters), "strings")
   expect_error(sym(1:2), "strings")